feat: move system Flatpak handling to reflect endlish-oesque method
This takes system Flatpak handling out of Yafti and does it the same way as we tried on endlish-oesque.
It adds an /etc/flatpak/install file with a list of Flatpaks to be installed, a step in the Containerfile
to read that file and install the apps without deploying (to preseed the needed files), a
ublue-flatpak-system-install script to read that file and install what's there without pulling remote
data, and a systemd unit to run that script at boot.